Re-organize into package and fix imports
