+++ udbg/__init__.py
@@ -1,0 +1,3 @@
+name = "udbg"
+
+from .udbg import *
