+++ udbg/modules/configs.py
@@ -32,8 +32,8 @@ import capstone
 import keystone
 from tabulate import tabulate
 
-import utils
-from modules.unicorndbgmodule import AbstractUnicornDbgModule
+import udbg.utils as utils
+from udbg.modules.unicorndbgmodule import AbstractUnicornDbgModule
 
 
 class Configs(AbstractUnicornDbgModule):
