+++ udbg/modules/find.py
@@ -31,8 +31,8 @@ import re
 
 from tabulate import tabulate
 
-import utils
-from modules.unicorndbgmodule import AbstractUnicornDbgModule
+import udbg.utils as utils
+from udbg.modules.unicorndbgmodule import AbstractUnicornDbgModule
 
 
 class Find(AbstractUnicornDbgModule):
