+++ udbg/modules/mappings.py
@@ -30,8 +30,8 @@
 
 from tabulate import tabulate
 
-import utils
-from modules.unicorndbgmodule import AbstractUnicornDbgModule
+import udbg.utils as utils
+from udbg.modules.unicorndbgmodule import AbstractUnicornDbgModule
 
 
 class Mappings(AbstractUnicornDbgModule):
