+++ udbg/modules/patches.py
@@ -29,8 +29,8 @@
 #############################################################################
 
 from tabulate import tabulate
-import utils
-from modules.unicorndbgmodule import AbstractUnicornDbgModule
+import udbg.utils as utils
+from udbg.modules.unicorndbgmodule import AbstractUnicornDbgModule
 
 
 class Patches(AbstractUnicornDbgModule):
